Add wavlake wallet (#144)
* fix avatar image

* clear firebase user on logout

* add loading to login/signup form

* fix nsec format

* fix random user name

* handle new /random response

* rm anon firebase login

* comment out iOS build

* fix nostr profile pic UI

* rm log

* fix icons

* bump versions

* add location message

* add enabled prob to chooser

* add setting for wavlake wallet

* add wavlake wallet to defaults check

* add send api call

* use send zap in zap hook

* add temp api call for ticket zap

* Update app/auth/nsec.tsx


* Update app/auth/nsec.tsx

* Update app/auth/nsec.tsx

* Update app/auth/nsec.tsx


* Update app/auth/nsec.tsx


* show some settings when not logged in

* update wavlake wallet zap logic

* add new nwc connection screen to login

* add lnurl to the nwc uri

* autoconnect nwc

* fix nwc responses

* rm use wavlake wallet setting

* rename auto-nwc page

* increase font size

* add ascAppId

* update readme

* add album deeplink to android

* handle secret key hex input

* add relay domain parsing

* bump build

* udpdate auto-nwc route

* add add-nwc redirect to second nsec login path

* bump build

* bump text size for nwc settings screen

* add accounting api env var

* bump build

* add device info lib

* update pw field

* fix nav path, adjust padding

* fix one tap zap bugs, parse nwc error responses

* bump build

* bump build

* remove default from nwc form, add required error messages

* improve auto-nwc form validation

* add env var for wallet pubkey

---------

+++ app/settings.tsx
@@ -1,10 +1,11 @@
-import { Button, Text, TextInput, WalletChooser } from "@/components";
+import { Text, TextInput, WalletChooser } from "@/components";
 import { useRouter } from "expo-router";
 import {
   Keyboard,
   TouchableWithoutFeedback,
   View,
   ScrollView,
+  Alert,
 } from "react-native";
 import { useEffect, useState } from "react";
 import { useAuth, useToast } from "@/hooks";
@@ -27,10 +28,19 @@ import { useSettings } from "@/hooks/useSettings";
 import { useSettingsQueryKey } from "@/hooks/useSettingsQueryKey";
 import { BUILD_NUM, VERSION } from "@/app.config";
 
+function getDomainFromWebSocket(wsAddress?: string) {
+  if (!wsAddress) return "NWC";
+  // Use a regular expression to match the domain
+  const match = wsAddress.match(/^wss?:\/\/([^/:]+)/i);
+
+  // Return the matched domain or null if no match
+  return match ? match[1] : "NWC";
+}
+
 export default function SettingsPage() {
   const toast = useToast();
   const router = useRouter();
-  const { pubkey } = useAuth();
+  const { pubkey, userIsLoggedIn } = useAuth();
   const { colors } = useTheme();
 
   const { data: settings } = useSettings();
@@ -45,6 +55,7 @@ export default function SettingsPage() {
     settings?.allowListeningActivity ?? false,
   );
   const [enableNWC, setEnableNWC] = useState(settings?.enableNWC ?? false);
+
   const [oneTapZap, setOneTapZap] = useState(settings?.oneTapZap ?? false);
   const [publishKind1, setPublishKind1] = useState(
     settings?.publishKind1 ?? false,
@@ -77,18 +88,36 @@ export default function SettingsPage() {
   };
 
   const onDeleteNWC = async () => {
-    pubkey && deleteNwcSecret(pubkey);
-    await cacheSettings(
-      {
-        nwcRelay: undefined,
-        nwcCommands: [],
-        nwcPubkey: undefined,
-        nwcLud16: undefined,
-        enableNWC: false,
-      },
-      pubkey,
+    Alert.alert(
+      "Are you sure you want to delete this connection?",
+      `You will no longer be able to send zaps with your ${getDomainFromWebSocket(
+        settings?.nwcRelay,
+      )} wallet. You will need to re-connect to use it again.`,
+      [
+        {
+          text: "Cancel",
+          style: "cancel",
+        },
+        {
+          text: "Delete Connection",
+          style: "destructive",
+          onPress: async () => {
+            pubkey && deleteNwcSecret(pubkey);
+            await cacheSettings(
+              {
+                nwcRelay: undefined,
+                nwcCommands: [],
+                nwcPubkey: undefined,
+                nwcLud16: undefined,
+                enableNWC: false,
+              },
+              pubkey,
+            );
+            queryClient.invalidateQueries(settingsKey);
+          },
+        },
+      ],
     );
-    queryClient.invalidateQueries(settingsKey);
   };
 
   // autosave settings on change
@@ -113,8 +142,6 @@ export default function SettingsPage() {
     publishKind1,
   ]);
 
-  if (!settings) return;
-
   return (
     <TouchableWithoutFeedback onPress={() => Keyboard.dismiss()}>
       <ScrollView
@@ -136,39 +163,40 @@ export default function SettingsPage() {
             selectedWallet={defaultZapWallet}
             onSelectedWalletChange={setDefaultZapWallet}
           />
-          <NWCSettings
-            nwcRelay={settings.nwcRelay}
-            enableNWC={enableNWC}
-            setEnableNWC={setEnableNWC}
-            onDeleteNWC={onDeleteNWC}
-            onAddNWC={onAddNWC}
-            nwcCommands={settings.nwcCommands}
-          />
-          <View
-            style={{
-              marginTop: 24,
-              marginBottom: 4,
-              flexDirection: "row",
-            }}
-          >
-            <View style={{ flex: 1 }}>
-              <Text bold>Listening activity</Text>
-              <Text>
-                Broadcast tracks you are listening to as a live status event to
-                Nostr relays.
-              </Text>
-            </View>
-            <Switch
-              value={allowListeningActivity}
-              onValueChange={setAllowListeningActivity}
-              color={brandColors.pink.DEFAULT}
-              trackColor={{
-                false: colors.border,
-                true: brandColors.pink.DEFAULT,
-              }}
-              thumbColor={colors.text}
+          {userIsLoggedIn && (
+            <NWCSettings
+              setEnableNWC={setEnableNWC}
+              onDeleteNWC={onDeleteNWC}
+              onAddNWC={onAddNWC}
             />
-          </View>
+          )}
+          {userIsLoggedIn && (
+            <View
+              style={{
+                marginTop: 24,
+                marginBottom: 4,
+                flexDirection: "row",
+              }}
+            >
+              <View style={{ flex: 1 }}>
+                <Text bold>Listening activity</Text>
+                <Text>
+                  Broadcast tracks you are listening to as a live status event
+                  to Nostr relays.
+                </Text>
+              </View>
+              <Switch
+                value={allowListeningActivity}
+                onValueChange={setAllowListeningActivity}
+                color={brandColors.pink.DEFAULT}
+                trackColor={{
+                  false: colors.border,
+                  true: brandColors.pink.DEFAULT,
+                }}
+                thumbColor={colors.text}
+              />
+            </View>
+          )}
           <View
             style={{
               marginTop: 24,
@@ -194,31 +222,33 @@ export default function SettingsPage() {
               thumbColor={colors.text}
             />
           </View>
-          <View
-            style={{
-              marginTop: 24,
-              marginBottom: 4,
-              flexDirection: "row",
-            }}
-          >
-            <View style={{ flex: 1 }}>
-              <Text bold>Publish comments to nostr</Text>
-              <Text>
-                Publish comments to your nostr feed. These comments will show up
-                in other nostr clients as kind 1 events.
-              </Text>
-            </View>
-            <Switch
-              value={publishKind1}
-              onValueChange={setPublishKind1}
-              color={brandColors.pink.DEFAULT}
-              trackColor={{
-                false: colors.border,
-                true: brandColors.pink.DEFAULT,
+          {userIsLoggedIn && (
+            <View
+              style={{
+                marginTop: 24,
+                marginBottom: 4,
+                flexDirection: "row",
               }}
-              thumbColor={colors.text}
-            />
-          </View>
+            >
+              <View style={{ flex: 1 }}>
+                <Text bold>Publish comments to nostr</Text>
+                <Text>
+                  Publish comments to your nostr feed. These comments will show
+                  up in other nostr clients as kind 1 events.
+                </Text>
+              </View>
+              <Switch
+                value={publishKind1}
+                onValueChange={setPublishKind1}
+                color={brandColors.pink.DEFAULT}
+                trackColor={{
+                  false: colors.border,
+                  true: brandColors.pink.DEFAULT,
+                }}
+                thumbColor={colors.text}
+              />
+            </View>
+          )}
         </View>
         <View style={{ flex: 1, alignSelf: "flex-start", marginBottom: 4 }}>
           <Text bold>Version information</Text>
@@ -232,23 +262,19 @@ export default function SettingsPage() {
 }
 
 const NWCSettings = ({
-  nwcRelay,
-  enableNWC,
   setEnableNWC,
   onDeleteNWC,
   onAddNWC,
-  nwcCommands,
 }: {
-  nwcRelay: string;
-  enableNWC: boolean;
   setEnableNWC: (enable: boolean) => void;
   onDeleteNWC: () => void;
   onAddNWC: () => void;
-  nwcCommands: string[];
 }) => {
+  const { data: settings } = useSettings();
+
   const { colors } = useTheme();
   const nwcCantPayInvoices =
-    !!nwcRelay && !nwcCommands.includes(payInvoiceCommand);
+    !!settings?.nwcRelay && !settings?.nwcCommands.includes(payInvoiceCommand);
 
   return (
     <View>
@@ -268,11 +294,13 @@ const NWCSettings = ({
               gap: 4,
             }}
           >
-            {nwcRelay && <CheckCircleIcon color={brandColors.mint.DEFAULT} />}
-            <Text>{nwcRelay || "Add a NWC compatible wallet."}</Text>
+            {settings?.nwcRelay && (
+              <CheckCircleIcon color={brandColors.mint.DEFAULT} />
+            )}
+            <Text>{settings?.nwcRelay || "Add a NWC compatible wallet."}</Text>
           </View>
         </View>
-        {nwcRelay ? (
+        {settings?.nwcRelay ? (
           <TrashIcon
             onPress={onDeleteNWC}
             color={brandColors.orange.DEFAULT}
@@ -296,7 +324,7 @@ const NWCSettings = ({
           wallet connection.
         </Text>
       )}
-      {nwcRelay && (
+      {settings?.nwcRelay && (
         <View
           style={{
             marginTop: 24,
@@ -307,13 +335,13 @@ const NWCSettings = ({
           <View style={{ flex: 1 }}>
             <Text bold>Enable NWC Wallet</Text>
             <Text>
-              {enableNWC
-                ? "Disable to use a different wallet for zaps."
-                : "Enable to use NWC wallet for zaps."}
+              {settings?.enableNWC
+                ? "Disable to use a different wallet."
+                : "Enable to use NWC wallet."}
             </Text>
           </View>
           <Switch
-            value={enableNWC}
+            value={settings?.enableNWC}
             onValueChange={setEnableNWC}
             disabled={nwcCantPayInvoices}
             color={brandColors.pink.DEFAULT}
